feat: add calculate_retry_time function

+++ app/settings.py
@@ -8,7 +8,7 @@ class Settings(BaseSettings):
 
     bot_token: str = "BOT_TOKEN"
     log_level: str = "INFO"
-    retry_time: int = 9  # seconds for another ticket finding retry
+    retry_time: float = 9.0  # seconds for another ticket finding retry
     max_concurrent_searches: int = 3
     request_timeout: float = 3.5
     retry_attempts: int = 7
